Extracted common test code into extension.
